pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "massgenotyping"
dynamic = ["version"]
description = 'Python package for microsatellite genotyping from amplicon sequencing data'
readme = "README.rst"
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = ["genotyping", "microsatellite", "NGS", "amplicon sequencing"]
authors = [
  { name = "kohyamat", email = "tetsuo.kohyama@gmil.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "biopython",
  "dataclasses-json",
  "fuzzysearch",
  "matplotlib",
  "natsort",
  "numpy",
  "pandas",
  "python-Levenshtein",
  "seaborn",
  "tqdm",
]

[project.scripts]
mgt = "massgenotyping:main"

[project.urls]
Documentation = "https://github.com/kohyamat/massgenotyping#readme"
Issues = "https://github.com/kohyamat/massgenotyping/issues"
Source = "https://github.com/kohyamat/massgenotyping"

[tool.hatch.version]
path = "src/massgenotyping/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/massgenotyping tests}"

[tool.coverage.run]
source_pkgs = ["massgenotyping", "tests"]
branch = true
parallel = true
omit = [
  "src/massgenotyping/__about__.py",
]

[tool.coverage.paths]
massgenotyping = ["src/massgenotyping", "*/massgenotyping/src/massgenotyping"]
tests = ["tests", "*/massgenotyping/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
